% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circClusterDE}
\alias{circClusterDE}
\title{circ-cluster DE function}
\usage{
circClusterDE(
  circObj,
  circ.cutoff = 2,
  DEmethod = c("Meta", "edgeR", "DESeq2")
)
}
\arguments{
\item{circObj}{updated circObj that contains individual circRNA DE results}

\item{circ.cutoff}{minimal circRNA number within the circ-cluster}

\item{DEmethod}{cluster DE methods: Meta-novel statistical method, edgeR, DESeq2}
}
\value{
circ-cluster DE result stored in list, A5BS cluster DE and A3BS cluster DE
}
\description{
circ-cluster DE function
}
