% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circNames}
\alias{circNames}
\title{handle column names of either CIRCexplorer2 or CIRI2}
\usage{
circNames(circ.method = c("CIRCexplorer2", "CIRI2"))
}
\arguments{
\item{circ.method}{specify circRNA calling method}
}
\value{
a vector of column names
}
\description{
handle column names of either CIRCexplorer2 or CIRI2
}
