% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circRNADE}
\alias{circRNADE}
\title{individual circRNA DE function}
\usage{
circRNADE(
  circObj,
  DEmethod = c("Pois", "GLM", "edgeR", "DESeq2"),
  formula_str = NULL
)
}
\arguments{
\item{circObj}{circObj contains circRNA.all.uniq and circRNA.all}

\item{DEmethod}{Pois method for small sample no covariates; GLM for large sample with covariates.}

\item{formula_str}{specify formular used for poisson GLM regression. e.g. "readNumber ~ condid + age + sex"}
}
\value{
updated circObj that contains inidividual circRNA DE result under circRNA.DE slot
}
\description{
individual circRNA DE function
}
