% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo.circs}
\alias{demo.circs}
\title{Example circRNA data from brain samples}
\format{
A GRanges object with circRNA coordinates and expression values
}
\source{
Processed from brain tissue samples
}
\usage{
demo.circs
}
\description{
A GRanges object containing circRNA expression data from brain samples
including metadata for differential expression analysis.
}
\keyword{datasets}
