% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCircCluster}
\alias{getCircCluster}
\title{clustering individual circRNA based on A5BS and A3BS events}
\usage{
getCircCluster(circObj)
}
\arguments{
\item{circObj}{circObj object}
}
\value{
a updated circObj that contains clustering information under slots A5BS.cluster and A3BS.cluster
}
\description{
clustering individual circRNA based on A5BS and A3BS events
}
