% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{getCircRNADE}
\alias{getCircRNADE}
\alias{getCircRNADE,circObj-method}
\title{Get individual circRNA differential expression results}
\usage{
getCircRNADE(x, ...)

\S4method{getCircRNADE}{circObj}(x)
}
\arguments{
\item{x}{circObj object}

\item{...}{Additional arguments}
}
\value{
data.frame with individual circRNA DE results
}
\description{
Get individual circRNA differential expression results
}
\examples{
\dontrun{
# Get all results
results <- getCircRNADE(circ.obj)
}
}
