% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{getClusterAnnotation}
\alias{getClusterAnnotation}
\title{Get cluster composition and annotation details}
\usage{
getClusterAnnotation(circObj, cluster_type, juncid)
}
\arguments{
\item{circObj}{circObj with cluster information}

\item{cluster_type}{Type of cluster ("A5BS" or "A3BS")}

\item{juncid}{Vector of junction IDs to get annotations for}
}
\value{
Data frame with circRNA-level annotation for specified clusters
}
\description{
Get cluster composition and annotation details
}
