% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{getClusterDE}
\alias{getClusterDE}
\title{Get cluster differential expression results}
\usage{
getClusterDE(circObj, cluster_type = "both")
}
\arguments{
\item{circObj}{circObj with completed cluster DE analysis}

\item{cluster_type}{Type of cluster to retrieve ("A5BS", "A3BS", or "both")}
}
\value{
Data frame with cluster DE results
}
\description{
Get cluster differential expression results
}
