% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makecircObj}
\alias{makecircObj}
\title{create S4 circObj object directly using output files from either CIRCexplorer2 or CIRI2}
\usage{
makecircObj(
  samplefiles,
  conditions = c(2, 2),
  circ.method = c("CIRCexplorer2", "CIRI2"),
  cutoff = 2
)
}
\arguments{
\item{samplefiles}{files path}

\item{conditions}{specify number of samples for each condition}

\item{circ.method}{specify method used for calling circRNA}

\item{cutoff}{minimal reads for circRNA, default is 2}
}
\value{
circObj S4 object
}
\description{
create S4 circObj object directly using output files from either CIRCexplorer2 or CIRI2
}
