% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makecircObjfromGRanges}
\alias{makecircObjfromGRanges}
\title{create S4 circObj object directly from circRNA GRanges}
\usage{
makecircObjfromGRanges(GRanges, sparse_filter = 0.05, metadata = NULL)
}
\arguments{
\item{GRanges}{circRNA in GRanges format}

\item{sparse_filter}{sparsity filtering of circRNA, based on proportion sample has the circRNA and the mean expression of circRNA.}

\item{metadata}{metadata information regarding the samples, covariates such as age, gender...}
}
\value{
circObj S4 object
}
\description{
create S4 circObj object directly from circRNA GRanges
}
