% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metainfo}
\alias{metainfo}
\title{Metadata for brain samples}
\format{
A data frame with sample information:
\describe{
  \item{sampleid}{Sample identifier}
  \item{condid}{Condition (1 = control, 2 = case)}
  \item{age}{Sample age}
  \item{sex}{Sample sex (M/F)}
}
}
\source{
Brain tissue sample collection
}
\usage{
metainfo
}
\description{
Sample metadata including condition, age, and sex information
for circRNA differential expression analysis.
}
\keyword{datasets}
