% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runPois.ztest}
\alias{runPois.ztest}
\title{individual circRNA DE method}
\usage{
runPois.ztest(dat)
}
\arguments{
\item{dat}{circRNA by sample counts matrix for circRNA}
}
\value{
DE results for individual circRNA
}
\description{
individual circRNA DE method
}
